Refactor API for defining derivatives.
