+++ app/models/document.rb
@@ -1,8 +1,5 @@
 class Document < GenericAsset
-  [:pdf_pages].each do |derivative|
-    delegate :"has_#{derivative}", :"has_#{derivative}=", :to => :workflow_metadata, :allow_nil => true
-    delegate :"#{derivative}_path", :"#{derivative}_path=", :to => :workflow_metadata, :allow_nil => true
-  end
+  has_derivatives :pdf_pages
 
   private
 
@@ -12,9 +9,5 @@ class Document < GenericAsset
 
   def derivative_class
     OregonDigital::Derivatives::Generators::DocumentDerivativeGenerator
-  end
-
-  def injector
-    @injector ||= OregonDigital.inject
   end
 end
