+++ app/models/image.rb
@@ -1,8 +1,5 @@
 class Image < GenericAsset
-  [:thumbnail, :medium, :pyramidal].each do |derivative|
-    delegate :"has_#{derivative}", :"has_#{derivative}=", :to => :workflow_metadata, :allow_nil => true
-    delegate :"#{derivative}_path", :"#{derivative}_path=", :to => :workflow_metadata, :allow_nil => true
-  end
+  has_derivatives :thumbnail, :medium, :pyramidal
 
   private
 
@@ -14,7 +11,4 @@ class Image < GenericAsset
     derivative_class.new(self, content, injector.thumbnail_runner(id), injector.medium_runner(id), injector.pyramidal_runner(id))
   end
 
-  def injector
-    @injector ||= OregonDigital.inject
-  end
 end
